GH-484 Add human readable error page
