+++ github-integration/src/main/java/de/zalando/zally/integration/ValidationController.kt
@@ -3,11 +3,15 @@ package de.zalando.zally.integration
 import de.zalando.zally.integration.config.logger
 import de.zalando.zally.integration.validation.ValidationService
 import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
+import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
+import javax.servlet.http.HttpServletRequest
 
 @RestController
 class ValidationController(private val validationService: ValidationService) {
@@ -31,6 +35,13 @@ class ValidationController(private val validationService: ValidationService) {
         validationService.validatePullRequest(payload, signature)
 
         log.info("Finished webhook processing")
+    }
+
+    @ExceptionHandler(Exception::class)
+    @ResponseBody
+    fun handleControllerException(request: HttpServletRequest, ex: Throwable): ResponseEntity<*> {
+        log.error("request failed", ex)
+        return ResponseEntity<Any>(HttpStatus.INTERNAL_SERVER_ERROR)
     }
 
 }
