+++ github-integration/src/main/java/de/zalando/zally/integration/reports/ReportService.kt
@@ -2,6 +2,7 @@ package de.zalando.zally.integration.reports
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import de.zalando.zally.integration.github.PullRequestEvent
+import de.zalando.zally.integration.validation.ValidationNotFoundException
 import de.zalando.zally.integration.validation.ValidationRepository
 import de.zalando.zally.integration.zally.ApiDefinitionResponse
 import org.springframework.stereotype.Service
@@ -11,7 +12,7 @@ class ReportService(private val validationRepository: ValidationRepository,
                     private val jsonObjectMapper: ObjectMapper) {
 
     fun getReport(id: Long): Report {
-        val validation = validationRepository.getOne(id)
+        val validation = validationRepository.findById(id).orElseThrow({ ValidationNotFoundException("Not found Validation with id $id") })
         val pullRequest = jsonObjectMapper.readValue(validation.pullRequestInfo, PullRequestEvent::class.java)
         val apiDefinitionResponse = jsonObjectMapper.readValue(validation.violations, ApiDefinitionResponse::class.java)
         return Report(validation.id, pullRequest, validation.apiDefinition, apiDefinitionResponse)
