+++ github-integration/src/main/java/de/zalando/zally/integration/validation/ValidationRepository.kt
@@ -1,5 +1,10 @@
 package de.zalando.zally.integration.validation
 
 import org.springframework.data.jpa.repository.JpaRepository
+import java.util.Optional
 
-interface ValidationRepository : JpaRepository<Validation, Long>
+interface ValidationRepository : JpaRepository<Validation, Long> {
+
+    fun findById(id: Long): Optional<Validation>
+
+}
