+++ github-integration/src/main/resources/templates/error.html
@@ -1,0 +1,38 @@
+<!DOCTYPE HTML>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>Error</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+    <style>
+        .error {
+            text-align: center;
+            font-weight: 400;
+            font-size: 3rem;
+            line-height: 3rem;
+            padding: 20px;
+        }
+
+        .details {
+            text-align: center;
+            font-weight: 300;
+            font-size: 2rem;
+            line-height: 2rem;
+            padding: 20px;
+        }
+
+        .exception {
+            margin: auto;
+            width: 50%;
+            padding: 20px;
+            text-align: center;
+        }
+    </style>
+</head>
+<body>
+
+<div class="error" th:text="${error}">Error title</div>
+<div class="details" th:text="${path}">/path/</div>
+<div class="exception" th:text="${exception}">bla foo</div>
+
+</body>
+</html>
